Reorganizando namespaces e disparando excessoes próprias
