+++ src/Categories/CategoriesException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Studos\Redacao1000\Categories;
+
+use Exception;
+
+final class CategoriesException extends Exception
+{
+}
