+++ src/Redaction/RedactionReport.php
@@ -1,7 +1,9 @@
 <?php
-namespace Studos\Redacao1000;
+
+namespace Studos\Redacao1000\Redaction;
 
 use DateTime;
+use Studos\Redacao1000\Base;
 
 final class RedactionReport extends Base
 {
