+++ src/Redaction/RedactionReportException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Studos\Redacao1000\Redaction;
+
+use Exception;
+
+class RedactionReportException extends Exception
+{
+}
