+++ src/Redaction/RedactionThemeException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Studos\Redacao1000\Authentication;
+
+use Exception;
+
+class RedactionThemeException extends Exception
+{
+}
