+++ src/School/SchoolClass.php
@@ -1,5 +1,8 @@
 <?php
-namespace Studos\Redacao1000;
+
+namespace Studos\Redacao1000\School;
+
+use Studos\Redacao1000\Base;
 
 final class SchoolClass extends Base
 {
