+++ src/School/SchoolException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Studos\Redacao1000\School;
+
+use Exception;
+
+class SchoolException extends Exception
+{
+}
