+++ src/Student/StudentException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Studos\Redacao1000\Student;
+
+use Exception;
+
+class StudentException extends Exception
+{
+}
