+++ tests/Redaction/Responses/reportFromPeriod.json
