+++ tests/School/Responses/subscribeClass.json
