+++ tests/School/Responses/subscribeClassWithInvalidCodeSchool.json
